Update event reminder functionality and Docker configuration
- Upgrade fakerphp/faker to version 1.24
- Add testReminder method to CsvEventReminderController for dispatching reminder emails
- Refactor EventReminder model to auto-generate event IDs
- Enhance CsvEventReminderService and EventReminderFactory for better event creation
- Introduce entrypoint.sh for improved Docker container startup
- Update Docker Compose files for better service management
- Modify README.md to reflect new features and setup instructions
