--- Apis.http
@@ -1,4 +1,0 @@
-### GET request to example server
-GET http://backend:8000/api/event-reminders
-
-###
