+++ api-service/Dockerfile
@@ -25,8 +25,9 @@ RUN chmod -R 777 storage bootstrap/cache
 # Add cron job for Laravel Scheduler
 RUN echo "* * * * * cd /app && php artisan schedule:run >> /var/log/laravel-cron.log 2>&1" | crontab -
 
-# Start cron in the background & process Laravel queues
-CMD service cron start && php artisan queue:work &
-CMD ["php", "artisan", "serve", "--host=0.0.0.0", "--port=8000"]
+COPY entrypoint.sh /entrypoint.sh
+RUN chmod +x /entrypoint.sh
 
 EXPOSE 8000
+
+ENTRYPOINT ["/entrypoint.sh"]
