+++ api-service/routes/web.php
@@ -3,12 +3,7 @@
 use App\Http\Controllers\Api\CsvEventReminderController;
 use App\Http\Controllers\Api\EventReminderController;
 use Illuminate\Foundation\Http\Middleware\VerifyCsrfToken;
-use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Route;
-
-//Route::get('/', function () {
-//    return view('welcome');
-//});
 
 Route::prefix('api')->group(function () {
     Route::prefix('event-reminders')->group(function () {
