+++ docker-compose.yml
@@ -9,14 +9,11 @@ services:
     network_mode: host
     volumes:
       - ./api-service:/app
-    environment:
-      DB_HOST: 127.0.0.1
-      DB_PORT: 3306
-      DB_DATABASE: event_db
-      DB_USERNAME: root
-      DB_PASSWORD: root
+    env_file:
+      - ./api-service/.env
     depends_on:
-      - db
+      db:
+        condition: service_healthy
 
   frontend:
     build:
@@ -37,8 +34,13 @@ services:
     environment:
       MYSQL_ROOT_PASSWORD: root
       MYSQL_DATABASE: event_db
-    volumes:
-      - db_data:/var/lib/mysql
+#    volumes:
+#      - db_data:/var/lib/mysql
+    healthcheck:
+      test: ["CMD", "mysqladmin", "ping", "-h", "localhost"]
+      interval: 10s
+      timeout: 5s
+      retries: 3
 
 volumes:
   db_data:
