+++ ui-service/src/api/types/eventType.ts
