+++ ui-service/src/app/providers.tsx
@@ -2,7 +2,7 @@
 
 import {ReactNode} from "react";
 import {Provider} from "react-redux";
-import {store} from "./store/store";
+import {store} from "@/store/store";
 
 export default function Providers({children}: { children: ReactNode }) {
     return <Provider store={store}>{children}</Provider>;
